Move controller_base into controllers/ directory.
