+++ whoc/controllers/hercules_wind_battery_controller.py
@@ -13,7 +13,7 @@
 # See https://nrel.github.io/wind-hybrid-open-controller for documentation
 
 import numpy as np
-from whoc.controller_base import ControllerBase
+from whoc.controllers.controller_base import ControllerBase
 
 
 class HerculesWindBatteryController(ControllerBase):
